feat(serde-sarif): add miri converter
