+++ .pre-commit-config.yaml
@@ -1,1 +1,1 @@
-/nix/store/vxglvm2j73n5w0ljrj372mrj922py4br-pre-commit-config.json
+/nix/store/860nhag5b16l86ylywx7n0jc56qqyib5-pre-commit-config.json
