+++ serde-sarif/src/converters/mod.rs
@@ -1,6 +1,13 @@
+#[cfg(any(feature = "clippy-converters", feature = "miri-converters"))]
+mod cargo;
+
 #[cfg(feature = "clippy-converters")]
 #[cfg_attr(doc, doc(cfg(feature = "clippy-converters")))]
 pub mod clippy;
+
+#[cfg(feature = "miri-converters")]
+#[cfg_attr(doc, doc(cfg(feature = "miri-converters")))]
+pub mod miri;
 
 #[cfg(feature = "hadolint-converters")]
 #[cfg_attr(doc, doc(cfg(feature = "hadolint-converters")))]
